+++ CHANGES.rst
@@ -4,12 +4,14 @@ Changes
 Unreleased
 ----------
 
+- Make NotSupportedError public. `#8 <https://github.com/dajiaji/pyseto/pull/8>`__
+
 Version 0.2.0
 -------------
 
 Released 2021-09-05
 
-- Add Token object as a response of decode(). `#165 <https://github.com/dajiaji/pyseto/pull/6>`__
+- Add Token object as a response of decode(). `#6 <https://github.com/dajiaji/pyseto/pull/6>`__
 
 Version 0.1.0
 -------------